requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
aiohttp==3.10.5
python-dotenv==1.0.1
streamlit==1.39.0
requests==2.32.3
pandas==2.2.3
plotly==5.24.1